i changed layout

+++ frontend/src/components/ProductList.tsx
@@ -1,4 +1,4 @@
-import { Link } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import { useCheckInputs } from "../context/useCheckInputs";
 import { useGetProducts } from "../hooks/useGetProducts";
 import { useDeleteProduct } from "../hooks/useDeleteProduct";
@@ -8,15 +8,41 @@ const ProductList = () => {
   const { products, isProductsPending } = useGetProducts();
   const { checkedProductIds, handleCheckboxChange } = useCheckInputs();
   const { deleteProduct, isDeleting } = useDeleteProduct();
+  const navigate = useNavigate();
 
   const handleDelete = () => {
-    deleteProduct(checkedProductIds);
+    if (checkedProductIds.length > 0) {
+      deleteProduct(checkedProductIds);
+    }
   };
 
   if (isProductsPending) return <p>Loading...</p>;
 
   return (
     <div className="w-full flex justify-center items-center flex-col py-4 lg:pt-32 lg:pb-10">
+      <header className="w-full h-[80px] flex items-center justify-center flex-col gap-2 fixed top-0 px-2 sm:px-4 md:px-6 lg:px-8 z-10 bg-slate-50 ">
+        <div className="w-full flex justify-between items-center px-4">
+          <h1 className="font-semibold text-lg sm:text-2xl lg:text-3xl">
+            Product List
+          </h1>
+          <div className="flex items-center gap-2">
+            <button
+              onClick={() => navigate("/productadd")}
+              className="py-1 px-3 border-[2px] border-b-4 border-r-4 border-slate-900 hover:border-r-[5px] hover:border-b-[5px] duration-100 transition-all"
+            >
+              ADD
+            </button>
+            <button
+              onClick={handleDelete}
+              id="delete-product-btn"
+              className="delete-checkbox py-1 px-3 border-[2px] border-b-4 border-r-4 border-slate-900 hover:border-r-[5px] hover:border-b-[5px] duration-100 transition-all"
+            >
+              {isDeleting ? "Deleting..." : "MASS DELETE"}
+            </button>
+          </div>
+        </div>
+        <div className="w-full flex justify-center items-center h-[2px] bg-black rounded-3xl "></div>
+      </header>
       <div className="w-full max-w-screen-xl bg-cover px-4 sm:px-6 lg:px-8">
         <div className="grid grid-cols-1 sm:grid-cols-2 md:grid-cols-3 lg:grid-cols-4 gap-6">
           {Array.isArray(products) && products.length > 0 ? (
@@ -46,24 +72,6 @@ const ProductList = () => {
           </p>
         </div>
       )}
-
-      {checkedProductIds.length > 0 && (
-        <div className="w-full flex justify-center items-center py-4">
-          <button
-            onClick={handleDelete}
-            className="py-2 px-4 bg-red-600 text-white font-semibold rounded-lg hover:bg-red-700 transition duration-150"
-            disabled={isDeleting}
-          >
-            {isDeleting ? "Deleting..." : "MASS DELETE"}
-          </button>
-        </div>
-      )}
-
-      <div className="w-full flex justify-center items-center py-4">
-        <Link to="/productadd" className="N">
-          ADD
-        </Link>
-      </div>
     </div>
   );
 };
